Revert dtype

+++ pennylane/devices/default_qubit.py
@@ -948,5 +948,5 @@ class DefaultQubit(QubitDevice):
         outcomes = outcomes[:, device_wires]
         recipes = recipes[:, device_wires]
 
-        # return self._cast(self._stack([outcomes, recipes]), dtype=np.uint8)
-        return self._cast(self._stack([outcomes, recipes]), dtype=np.float32)
+        return self._cast(self._stack([outcomes, recipes]), dtype=np.uint8)
+        # return self._cast(self._stack([outcomes, recipes]), dtype=np.float32)
